+++ src/pvenv/scripts/pvenv.sh
@@ -1,0 +1,15 @@
+function venv {
+    case "$1" in
+    list)
+        lsvenv "${@:2}"
+        ;;
+    *)
+        pvenv "$@"
+        return
+        ;;
+    esac
+}
+
+function lsvenv {
+    pvenv list "$@"
+}