+++ src/pvenv/subcommands/__init__.py
@@ -1,0 +1,3 @@
+from pvenv.subcommands import lsvenv
+
+__all__ = ["lsvenv"]